refactor: moved registry and component_factory to joint module
